Changed price and rating columns to allow for decimals

+++ database/migrations/2023_04_24_065529_create_products_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('description');
             $table->string('image');
-            $table->integer('price');
-            $table->integer('rating');
+            $table->decimal('price', 10, 2);
+            $table->decimal('rating', 2, 2);
             $table->timestamps();
         });
     }
